pyproject.toml
[project]
name = "1-chat-with-mysql-database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "mysql-connector-python>=9.6.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy",
    "streamlit>=1.55.0",
]
